chore: move 'devworkspace-client' -> 'devworkspaceClient'
