+++ packages/dashboard-backend/jest.config.js
@@ -19,9 +19,11 @@ module.exports = {
   moduleDirectories: [
     'node_modules',
   ],
+  coveragePathIgnorePatterns: [
+    'src/devworkspaceClient/__tests__/*',
+  ],
   testPathIgnorePatterns: [
-    '/node_modules/',
-    'src/devworkspace-client/__tests__/errors.spec.ts',
-    'src/devworkspace-client/__tests__/integration.spec.ts',
+    'src/devworkspaceClient/__tests__/errors.spec.ts',
+    'src/devworkspaceClient/__tests__/integration.spec.ts',
   ]
 };
