+++ packages/dashboard-backend/src/devworkspaceClient/__tests__/fixtures/sample-devfile-plugins.yaml
